refactor: add InstallDependencies action
